+++ tests/projects/test_python.py
@@ -17,22 +17,6 @@ def test_python_template(cookies):
     assert (result.project + "/dev-requirements.txt").exists()
 
 
-def test_python_template_pipenv(cookies):
-    result = cookies.bake(
-        template=f"{os.path.dirname(os.path.abspath(__file__))}/../../project/python",
-        extra_context={"python_package_management": "pipenv"},
-    )
-    assert 0 == result.exit_code
-    assert result.exception is None
-    assert result.project.isdir()
-    assert (result.project + "/Pipfile").exists()
-    assert (result.project + "/Pipfile.lock").exists()
-    assert not (result.project + "/requirements.in").exists()
-    assert not (result.project + "/requirements.txt").exists()
-    assert not (result.project + "/dev-requirements.in").exists()
-    assert not (result.project + "/dev-requirements.txt").exists()
-
-
 def test_python_template_no_role(cookies):
     result = cookies.bake(
         template=f"{os.path.dirname(os.path.abspath(__file__))}/../../project/python",